move configuration checking to config manager
